@@ -2,11 +2,14 @@
 
 namespace App\Entity;
 
+use ApiPlatform\Core\Annotation\ApiFilter;
 use ApiPlatform\Core\Annotation\ApiResource;
 use Doctrine\ORM\Mapping as ORM;
+use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\SearchFilter;
 
 /**
  * @ApiResource()
+ * @ApiFilter(SearchFilter::class, properties={"completed": "exact"})
  * @ORM\Entity(repositoryClass="App\Repository\ItemRepository")
  */
 class Item
@@ -16,12 +19,17 @@ class Item
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
      */
-    private $id;
+    private int $id;
 
     /**
      * @ORM\Column(type="string", length=255)
      */
-    private $name;
+    private string $name;
+
+    /**
+     * @ORM\Column(type="boolean")
+     */
+    private bool $completed = false;
 
     public function getId(): ?int
     {
@@ -38,5 +46,15 @@ class Item
         $this->name = $name;
 
         return $this;
+    }
+
+    public function isCompleted(): bool
+    {
+        return $this->completed;
+    }
+
+    public function setCompleted(bool $completed): void
+    {
+        $this->completed = $completed;
     }
 }
